changement de type dans getTemp

+++ LocationApp/src/main/java/fr/univtours/polytech/locationapp/dao/TempDao.java
@@ -1,5 +1,5 @@
 package fr.univtours.polytech.locationapp.dao;
 
 public interface TempDao {
-	public Double getTemp(int lon, int lat);
+	public Double getTemp(Double lon, Double lat);
 }
